+++ docs/content/misc/release_notes.rst
@@ -4,6 +4,7 @@ Release notes
 Version 0.4.0
 -------------
 - Created new version of the Matlab API.
+- Added simplified TGC control through tgcStart and tgcSlope parameters.
 
 Version 0.3.0
 -------------